Update API, Readme, etc (#1)
* Explicitly enable more Tyler tile downloads

* Refactor types and functions to use Float64 for precision; update README for clarity; remove unused dependencies from Project.toml; enhance error handling in match and route services.

+++ src/functions.jl
@@ -1,5 +1,5 @@
 """
-    distance(response, ...) -> Float32
+    distance(response, ...) -> Float64
 
 Compute distance from OSRM response objects. Methods are defined for:
 - `RouteResponse`: returns total route distance
@@ -9,7 +9,7 @@ Compute distance from OSRM response objects. Methods are defined for:
 function distance end
 
 """
-    duration(response, ...) -> Float32
+    duration(response, ...) -> Float64
 
 Compute duration from OSRM response objects. Methods are defined for:
 - `RouteResponse`: returns total route duration
